+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Added
 ^^^^^
 
 - ``dtool.DataSet.empty_overlay`` method
+- ``dtool.DataSet.persist_overlay`` method
 - ``dtool info`` CLI command
 - ``dtool.clickutils.info_from_path`` function
 - ``dtool.DataSet.item_path_from_hash`` method